Fix integer conversion
The io_data->nb_backup_sectors field is an uint32_t, but when compared
to sector_offset (which is an off_t), sector_offset was cast to
uint64_t. The cast is now fixed to uint32_t.

+++ src/inouts/sectors.c
@@ -344,7 +344,7 @@ static void* thread_decrypt(void* params)
 
 		/* Check for sectors fixing and non-encrypted sectors */
 		if(version == V_SEVEN &&
-		   (uint64_t)sector_offset < io_data->nb_backup_sectors)
+		   (uint32_t)sector_offset < io_data->nb_backup_sectors)
 		{
 			/*
 			 * The firsts sectors are encrypted in a different place on a
